+++ config/Config.ts
@@ -1,0 +1,18 @@
+export type Config = {
+  contractRegistry: {
+    address: string
+  },
+  token: {
+    totalSupply: string,
+    name: string,
+    symbol: string
+  },
+  chain: {
+    blockPadding: number
+  },
+  validators: [
+    {
+      location: string
+    }
+  ]
+};