+++ config/production.ts
@@ -1,4 +1,6 @@
-const production = {
+import {Config} from './Config';
+
+const production: Config = {
   contractRegistry: {
     address: '0x8d16D5D2859f4c54b226180A46F26D57A4d727A0'
   },