+++ scripts/deploy-all.ts
@@ -10,11 +10,11 @@ async function main() {
   if (isLocalNetwork()) {
     console.log('registering contracts...');
     await deployAllContracts(registry.address, true);
-    //await registry.importContracts(Object.values(addresses));
-
     console.log('...done - local network ready.');
   } else {
-    console.log(`setup registry address in config file and run 'npm run deploy:contracts:${process.env.NODE_ENV}'`);
+    console.log('\n⚠️⚠️⚠️⚠️⚠️\n',
+      `setup registry address in config file and run\n\n> npm run deploy:contracts:${process.env.NODE_ENV}`,
+      '\n\nto finish deployment\n⚠️⚠️⚠️⚠️⚠️\n');
   }
 
   console.log('Registry:', registry.address);