+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2021-03-10
+### Added
+- individual script for deploy/update `Chain`
+- getters that allow access `Block` data for external calls
+
+### Fixed
+- new contract uses offset for block count, so we can preserve `blockHeight`
+
 ## [0.1.0] - 2021-02-25
 ### Added
 - initial version